Fix the bug of finding try blocks.
In former version, variable "end" may be null, which results in wrong closure of try blocks.
And rather than finding the closure, get the blocks in bytecode order is a more natural and correct way which is closer to JVM behavior.
But due to the mess of relation between labelNode and blocks, there might exist some subtle bugs. So more tests are needed.